Refactor core yubikit and ykman classes.
- Moved YUBIKEY, PID, and USB_INTERFACE into yubikit.core and changed
  the methods on these.
- Moved read_info, get_name and is_fips_version into yubikit.support.
