+++ ykman/base.py
@@ -25,56 +25,8 @@
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
 
-from yubikit.core import TRANSPORT, YubiKeyDevice
-from yubikit.management import USB_INTERFACE
-from enum import Enum, IntEnum, unique
+from yubikit.core import TRANSPORT, PID, YubiKeyDevice
 from typing import Optional, Hashable
-
-
-@unique
-class YUBIKEY(Enum):
-    """YubiKey hardware platforms."""
-
-    YKS = "YubiKey Standard"
-    NEO = "YubiKey NEO"
-    SKY = "Security Key by Yubico"
-    YKP = "YubiKey Plus"
-    YK4 = "YubiKey"  # This includes YubiKey 5
-
-    def get_pid(self, interfaces: USB_INTERFACE) -> "PID":
-        suffix = "_".join(
-            t.name or str(t) for t in USB_INTERFACE if t in USB_INTERFACE(interfaces)
-        )
-        return PID[self.name + "_" + suffix]
-
-
-@unique
-class PID(IntEnum):
-    """USB Product ID values for YubiKey devices."""
-
-    YKS_OTP = 0x0010
-    NEO_OTP = 0x0110
-    NEO_OTP_CCID = 0x0111
-    NEO_CCID = 0x0112
-    NEO_FIDO = 0x0113
-    NEO_OTP_FIDO = 0x0114
-    NEO_FIDO_CCID = 0x0115
-    NEO_OTP_FIDO_CCID = 0x0116
-    SKY_FIDO = 0x0120
-    YK4_OTP = 0x0401
-    YK4_FIDO = 0x0402
-    YK4_OTP_FIDO = 0x0403
-    YK4_CCID = 0x0404
-    YK4_OTP_CCID = 0x0405
-    YK4_FIDO_CCID = 0x0406
-    YK4_OTP_FIDO_CCID = 0x0407
-    YKP_OTP_FIDO = 0x0410
-
-    def get_type(self) -> YUBIKEY:
-        return YUBIKEY[self.name.split("_", 1)[0]]
-
-    def get_interfaces(self) -> USB_INTERFACE:
-        return USB_INTERFACE(sum(USB_INTERFACE[x] for x in self.name.split("_")[1:]))
 
 
 class YkmanDevice(YubiKeyDevice):
